+++ test/lein_viz/sample.clj
@@ -13,12 +13,12 @@
 (defn make-graph
   []
   {:graph {:service [:db :mailer]
-           :db {:datasource :conn-pool}
-           :datasource {:db-host :config
-                        :db-port :config
-                        :database :config
-                        :username :config
-                        :password :config}
+           :db {:conn-pool :datasource}
+           :datasource [:db-host
+                        :db-port
+                        :database
+                        :username
+                        :password]
            :mailer [:smtp-host :smtp-port]}
    :seed [:db-host :db-port :username :password]})