+++ CHANGES.md
@@ -18,8 +18,8 @@
 * Visualization
   * Support for labeled edges via dependency map of edge-labels to adjacency-list
 * CLI support
-  * Support for choosing selector `["-e" "--selector SELECTOR" "Selector in project.clj"]`
-  * BREAKING CHANGE: Target is now called source `["-s" "--source SOURCE" "Fully qualified defn var name (arity-0)"]`
+  * BREAKING CHANGE: Target is now called selector: `["-e" "--selector SELECTOR" "Selector in project.clj"]`
+  * Source may be specified via command-line: `["-s" "--source SOURCE" "Fully qualified defn var name (arity-0)"]`
   * BREAKING CHANGE: Source may now be placed only under the `:source` key under a selector
   * Read payload from STDIN when source is `:stdin`
   * Flag to show/hide missing nodes `["-m" "--hide-missing" "Hide missing nodes in a graph"]`